add fields in result_page.dart

+++ lib/calculator_brain.dart
@@ -4,7 +4,7 @@ class CalculatorBrain {
   CalculatorBrain({this.weight, this.height});
 
   final int weight;
-  final int height;
+  final double height;
   double _bmi;
 
   String calculateBMI() {
